Base code for testing hash table functions
A structure is created for the hash table, as well as for the items.
Functions are implemented to create a hash table and an item.
A function is implemented to compute the hash.
Functions are implemented to insert, delete, and search in the hash table.
A function is implemented to delete the hash table.